requirements.txt
apache-airflow==2.7.1
psycopg2-binary==2.9.9